refactor: move Loader
